% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno.R
\name{anno_group}
\alias{anno_group}
\title{Processes an Arrow table of laboratory data, performing several transformations
and annotations to prepare the data for analysis. This includes converting the
Arrow table to a DuckDB table, ordering the data, grouping entries, and removing
unnecessary columns.}
\usage{
anno_group(arrow_lab)
}
\arguments{
\item{arrow_lab}{An Arrow table containing laboratory data. This table is expected
to have certain columns (\code{CKSUM}, \code{TermLine}, \code{Term}, etc.) necessary for the
processing steps within the function.}
}
\value{
Returns a DuckDB table after applying a series of transformations. The
transformations include ordering by \code{CKSUM} and \code{TermLine}, creating entry groups
based on the \code{Term} column values, and removing the \code{Source} column. The table
is further processed with DuckDB-specific computations before being returned.
}
\description{
Processes an Arrow table of laboratory data, performing several transformations
and annotations to prepare the data for analysis. This includes converting the
Arrow table to a DuckDB table, ordering the data, grouping entries, and removing
unnecessary columns.
}
\details{
The function starts by converting the input Arrow table to a DuckDB table
for efficient processing. It then applies a window order based on \code{CKSUM} and
\code{TermLine} columns. Grouping is performed by creating entry groups where each
group starts with a row where \code{Term} is "ACC". The \code{Source} column is removed
from the final output, and the table is computed in DuckDB before being returned.
}
