% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{create_project}
\alias{create_project}
\title{Create Project Directory Structure}
\usage{
create_project(base_dir = "raw", copy_files = FALSE)
}
\arguments{
\item{base_dir}{The base directory relative to the current working directory.
Defaults to "raw".}

\item{copy_files}{Logical flag to indicate whether specific files should be copied
from a source directory to their respective new subdirectories. Defaults to FALSE.}
}
\value{
Invisible NULL. This function is primarily used for its side effects of creating
directories and optionally copying files.
}
\description{
Sets up a project directory structure in a specified base directory. Optionally,
copies predefined files to their respective subdirectories.
}
