% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ddb_col_hash}
\alias{ddb_col_hash}
\title{Generate a Hashed Column in a DuckDB Table}
\usage{
ddb_col_hash(ddb, col, len = 6, salt = NULL, new_name = NULL)
}
\arguments{
\item{ddb}{A DuckDB connection or table.}

\item{col}{The name of the column to hash.}

\item{len}{The length of the hash to be used. Defaults to 6.}

\item{salt}{An optional salt to be appended to the column values before hashing. Defaults to NULL.}

\item{new_name}{An optional new name for the hashed column. Defaults to the original column name with "_h" appended.}
}
\value{
A modified DuckDB table with the hashed column added.
}
\description{
This function adds a new column to a DuckDB table with hashed values based on an existing column.
}
