% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ddb_col_redact}
\alias{ddb_col_redact}
\title{Redact Sensitive Information in a DuckDB Table Column}
\usage{
ddb_col_redact(ddb, col, new_name = NULL)
}
\arguments{
\item{ddb}{A DuckDB connection or table.}

\item{col}{The name of the column containing sensitive information to be redacted.}

\item{new_name}{An optional new name for the redacted column. Defaults to the original column name with "_r" appended.}
}
\value{
A modified DuckDB table with the redacted column added.
}
\description{
This function redacts sensitive information in a specified column of a DuckDB table. It replaces dates, account numbers, and MRNs with a placeholder.
}
