% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ddb_col_shift}
\alias{ddb_col_shift}
\title{Shift Dates in a DuckDB Table Based on Calculated Shifts}
\usage{
ddb_col_shift(ddb, col, new_name = NULL)
}
\arguments{
\item{ddb}{A DuckDB connection or table.}

\item{col}{The name of the column containing dates to be shifted.}

\item{new_name}{An optional new name for the shifted date column. Defaults to the original column name with "_s" appended.}
}
\value{
A modified DuckDB table with the date-shifted column added.
}
\description{
This function shifts dates in a specified column of a DuckDB table based on pre-calculated date shifts.
}
