% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ddb_ds_get}
\alias{ddb_ds_get}
\title{Calculate Date Shifts Based on MRN in a DuckDB Table}
\usage{
ddb_ds_get(ddb, salt = NULL)
}
\arguments{
\item{ddb}{A DuckDB connection or table.}

\item{salt}{An optional salt to be appended to the MRN values before calculating the shift. Defaults to NULL.}
}
\value{
A modified DuckDB table with the date shift column added.
}
\description{
This function adds a date shift column to a DuckDB table based on the modulus of the MRN (Medical Record Number).
}
