% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deid.R
\name{deid_acc}
\alias{deid_acc}
\title{Deid Accession Files}
\usage{
deid_acc(arrow, crosswalk)
}
\arguments{
\item{arrow}{coherently parsed and PHI data}

\item{crosswalk}{crosswalk containing identifying mappings}
}
\description{
Deid Accession Files
}
