% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deid.R
\name{deid_bat}
\alias{deid_bat}
\title{Deid Battery Files}
\usage{
deid_bat(arrow, crosswalk)
}
\arguments{
\item{arrow}{coherently parsed and PHI data}

\item{crosswalk}{crosswalk containing identifying mappings}
}
\description{
Deid Battery Files
}
