% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deid.R
\name{deid_body}
\alias{deid_body}
\title{De-identify and Parse Body Data}
\usage{
deid_body(anno_path, cw_path, type = "lab")
}
\arguments{
\item{anno_path}{The path to the annotation file.}

\item{cw_path}{The path to the codebook file.}

\item{type}{The type of data to process, either \code{"lab"} for laboratory data or \code{"icd"} for ICD data. Default is \code{"lab"}.}
}
\value{
A list containing the de-identified and parsed data.
}
\description{
This function de-identifies and parses body data based on the specified type.
It handles laboratory data (\code{lab}) and ICD data (\code{icd}).
}
