% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deid.R
\name{deid_head}
\alias{deid_head}
\title{Deid Head Files}
\usage{
deid_head(arrow, save_crosswalk = TRUE)
}
\arguments{
\item{arrow}{coherently parsed and PHI data}

\item{save_crosswalk}{option to save crosswalk containing identifying mappings}
}
\description{
Deid Head Files
}
