% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deid.R
\name{deid_lab_body_meta}
\alias{deid_lab_body_meta}
\title{Deid Body Metadata of Files}
\usage{
deid_lab_body_meta(arrow, crosswalk)
}
\arguments{
\item{arrow}{coherently parsed and PHI data}

\item{crosswalk}{crosswalk containing identifying mappings}
}
\description{
Unlike other deid_**, here we parse first to get keys to join on.
}
