% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{main_pipeline}
\alias{main_pipeline}
\title{Main Data Processing Pipeline}
\usage{
main_pipeline(
  dir,
  type = "raw",
  data_type = "labs",
  clean = TRUE,
  save = c(NULL, "arrow", "csv"),
  salt = 123
)
}
\arguments{
\item{dir}{A character string specifying the directory containing the data to be processed.}

\item{type}{A character string indicating the type of data to be processed. Default is "raw".}

\item{data_type}{A character string specifying the type of data. Default is "labs".}

\item{clean}{A logical value indicating whether to clean up intermediate files. Default is TRUE.}

\item{save}{A character vector specifying the format to save the processed data. Options are "arrow" and "csv". Default is NULL.}

\item{salt}{A numeric value used for anonymization purposes. Default is 123.}
}
\value{
A list containing file paths of saved processed data.
}
\description{
This function serves as a wrapper around several internal functions to process and manage data.
It handles raw data parsing, cleaning, anonymization, and saving of processed data in various formats.
}
