% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mart.R
\name{mart_lb}
\alias{mart_lb}
\title{Join and Process Laboratory Data}
\usage{
mart_lb(duck, acc, bat, head)
}
\arguments{
\item{duck}{A DuckDB object containing laboratory data.}

\item{acc}{A data frame or tibble containing de-identified accession data.}

\item{head}{A data frame or tibble containing de-identified header data.}
}
\value{
A data frame or tibble with selected and processed laboratory data.
}
\description{
This function joins de-identified accession and header data with a DuckDB object containing laboratory data,
selects specific columns, and processes the data.
}
