% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mart.R
\name{mart_lc_sens}
\alias{mart_lc_sens}
\title{Join and present prepared Sensitivity Laboratory Data}
\usage{
mart_lc_sens(duck, acc, bat, head, .type = NULL)
}
\arguments{
\item{duck}{A DuckDB object containing laboratory data.}

\item{acc}{A data frame or tibble containing de-identified accession data.}

\item{bat}{A data frame or tibble containing de-identified accession bat.}

\item{head}{A data frame or tibble containing de-identified header data.}

\item{.type}{Option for non-standard data processing. Only RD supported.}
}
\value{
A data frame or tibble with selected and processed laboratory data.
}
\description{
Join and present prepared Sensitivity Laboratory Data
}
