% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_comma}
\alias{parse_comma}
\title{Parse Comma-Separated Values in a Column}
\usage{
parse_comma(data, col)
}
\arguments{
\item{data}{A data frame or tibble.}

\item{col}{The name of the column to be parsed.}
}
\value{
A data frame or tibble with new columns created from the comma-separated values.
}
\description{
This function parses a specified column in a data frame, splitting its values by commas
and creating new columns for each part.
}
