% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{parse_csv_to_arrow}
\alias{parse_csv_to_arrow}
\title{Parses preformatted files from a data manager.}
\usage{
parse_csv_to_arrow(base_dir = "raw", sub_dir)
}
\description{
Parses preformatted files from a data manager.
}
