% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_lb_base}
\alias{parse_lb_base}
\title{Parse Laboratory Data}
\usage{
parse_lb_base(lab_dat_base)
}
\arguments{
\item{lab_dat_base}{A dataframe containing the base laboratory data with a column named 'Description'.}
}
\value{
A list containing three dataframes: \code{char} for character values, \code{num} for numeric values, and \code{err} for erroneous values.
}
\description{
This function parses laboratory data, separating descriptions into multiple components
and further processing the values into numeric and character categories.
}
