% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{parse_out}
\alias{parse_out}
\title{Execute Parsing Script}
\usage{
parse_out(base_dir = "raw")
}
\arguments{
\item{base_dir}{The base directory where the script will perform its operations.
Defaults to "raw".}
}
\value{
Invisible NULL. This function is primarily used for its side effect of executing
a shell script.
}
\description{
Executes a shell script (\code{parse.sh}) located in the \code{extdata} directory of the
\code{marsrap} package. This script is intended to parse and process files in a specified
base directory.
}
\examples{
\dontrun{
  parse_out(base_dir = "project_data")
}
}
