% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_outtxt}
\alias{read_outtxt}
\title{Read Specific Sections from a Text File}
\usage{
read_outtxt(out.txt, type)
}
\arguments{
\item{out.txt}{The path to the input text file.}

\item{type}{The type of section to read, one of "head", "acc", "bat", "body".}
}
\value{
A tibble corresponding to the section read, with appropriate column names and modifications.
}
\description{
This function reads different sections from a text file based on the section type.
It uses custom column names for each section type.
}
