% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{save_duck}
\alias{save_duck}
\title{Save Data in Specified Format}
\usage{
save_duck(duck, dir_name = "mart", file_name, format = c("arrow", "csv"))
}
\arguments{
\item{duck}{The dataset to be saved.}

\item{dir_name}{A character string specifying the directory where the file will be saved. Default is "mart".}

\item{file_name}{A character string specifying the name of the file to be saved.}

\item{format}{A character vector specifying the format to save the file. Options are "arrow" and "csv". Default is "arrow".}
}
\value{
A character string representing the file path of the saved file.
}
\description{
This function saves a given dataset (duck) into a specified directory and file format (either "arrow" or "csv").
}
