% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{separate_duckdb}
\alias{separate_duckdb}
\title{Separate Delimited Values into Multiple Columns}
\usage{
separate_duckdb(data, col, into, sep)
}
\arguments{
\item{data}{A data frame or tibble.}

\item{col}{The name of the column to be parsed.}

\item{into}{A vector of column names to be created from the split values.}

\item{sep}{The delimiter used to split the values.}
}
\value{
A data frame or tibble with new columns created from the delimited values.
}
\description{
This function separates a specified column in a data frame, splitting its values by a given delimiter
and creating new columns for each part.
}
